load-metrics supports a custom transform on the metric values.
The metric-defs file can have an optional fifth column, that gives the
path to a python script used to transform the value.  The script should have one
method with signature 'transform(group, category, name, value)' and returns a
transformed value.  The parameters passed to 'transform' are as follows:
  - group: gets the metric group
  - category: gets the metric category, or None if not defined
  - name: gets the metric name
  - value: gets the metric value
If the script is not found at the given path, we attempt to find it in the same
directory as the metrics definition file.

This may be useful if the metric name sames "pct" or "percentage" but really it
calculates a rate (I'm looking at you Picard metrics).